fix: 최초 로그인 유저 auth 문제 수정
